Moving to own file and error handling cleanup.
